+++ dbms/src/Common/typeid_cast.h
@@ -6,7 +6,7 @@
 #include <string>
 
 #include <Common/Exception.h>
-#include <Common/demangle.h>
+#include <common/demangle.h>
 
 
 namespace DB
