+++ libs/libcommon/src/tests/CMakeLists.txt
@@ -20,3 +20,6 @@ add_check(multi_version)
 add_executable (unit_tests_libcommon gtest_json_test.cpp gtest_strong_typedef.cpp)
 target_link_libraries (unit_tests_libcommon gtest_main common)
 add_check(unit_tests_libcommon)
+
+add_executable (dump_variable dump_variable.cpp)
+target_link_libraries (dump_variable clickhouse_common_io)
