+++ libs/libcommon/src/tests/dump_variable.cpp
@@ -1,4 +1,4 @@
-#include <Common/iostream_debug_helpers.h>
+#include <common/iostream_debug_helpers.h>
 
 #include <iostream>
 #include <memory>
